Agregada autenticación con JWT

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find({});
+    const goals = await Goal.find({ user: req.user._id });
 
     res.status(200).json(goals)
 });
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('No goal text');
     };
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id
     })
     res.status(200).json(goal)
 });
@@ -37,6 +39,18 @@ const updateGoals = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('No goal found')
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if(!user) {
+        res.status(401)
+        throw new Error('No user found')
+    }
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Not authorized to update goal')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -58,6 +72,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
         throw new Error('No goal found')
     }
 
+    const user = await User.findById(req.user.id);
+
+    if(!user) {
+        res.status(401)
+        throw new Error('No user found')
+    }
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Not authorized to update goal')
+    }
+    
     await Goal.findByIdAndDelete(req.params.id);
 
     res.status(200).json({ id: req.params.id })
